Setup for lexing part
