+++ calcli/core/CMakeLists.txt
@@ -1,0 +1,31 @@
+# Lists of hearders and sources
+set(calcli-core-headers
+	lexer.hpp
+)
+
+set(calcli-core-sources
+	lexer.cpp
+)
+
+# Executable compilation
+set(lib_name calcli-core)
+add_library(${lib_name} STATIC ${calci-core-hearders} ${calcli-core-sources})
+
+target_include_directories(${lib_name} PRIVATE ${CMAKE_SOURCE_DIR})
+
+target_compile_features(${lib_name} PRIVATE cxx_std_17)
+
+target_compile_options(${lib_name} PRIVATE 
+	"-Wall"
+	"-Wextra"
+	"-pedantic"
+)
+
+include(${CMAKE_SOURCE_DIR}/cmake/OutputDirectories.cmake)
+make_lib_directory(lib_directory)
+
+set_target_properties(${lib_name} PROPERTIES 
+	LINKER_LANGUAGE CXX
+	CXX_EXTENSIONS OFF
+	ARCHIVE_OUTPUT_DIRECTORY ${lib_directory}
+)
