+++ tests/CMakeLists.txt
@@ -1,7 +1,7 @@
 set(test_calcli_exe test-calcli)
 
 set(test_calcli_sources 
-	test_toto.cpp
+	test_lexer.cpp
 	test_calcli.cpp
 )
 
@@ -9,15 +9,23 @@ add_executable(${test_calcli_exe} ${test_calcli_sources})
 
 target_include_directories(${test_calcli_exe} PRIVATE ${CMAKE_SOURCE_DIR})
 
+target_compile_features(${test_calcli_exe} PRIVATE cxx_std_17)
+
+target_compile_options(${test_calcli_exe} PRIVATE 
+	"-Wall"
+	"-Wextra"
+	"-pedantic"
+)
+
+include(${CMAKE_SOURCE_DIR}/cmake/OutputDirectories.cmake)
+make_lib_directory(lib_directory)
+make_bin_directory(bin_directory)
+
 target_link_libraries(${test_calcli_exe}
+	${lib_directory}/libcalcli-core.a
 	gtest gtest_main
 	stdc++fs
 )
-
-target_compile_features(${test_calcli_exe} PRIVATE cxx_std_17)
-
-include(${CMAKE_SOURCE_DIR}/cmake/OutputDirectories.cmake)
-make_bin_directory(bin_directory)
 
 set_target_properties(${test_calcli_exe} PROPERTIES 
 	LINKER_LANGUAGE CXX
