--- tests/test_toto.cpp
@@ -1,9 +1,0 @@
-#include <gtest/gtest.h>
-
-
-TEST(test_toto, toto)
-{
-	const int value = 1;
-	const int expected = 1;
-	EXPECT_TRUE(value == expected);
-}
